refactor page editor
